requirements.txt
Django>=5.0,<6.0
djangorestframework
django-select2
mysqlclient
